feat(wezterm): adjust opacity and blur settings
The window_background_opacity has been reduced from 0.92 to 0.9 and
macos_window_background_blur has been increased from 30 to 50. The
background color for the active tab has been changed to a darker shade
and the opacity of the inactive tabs has been reduced. The function
for updating the right status has been removed.